Fix size_t/uint64_t types (#586)
* Remove uint32 in favor of uint32_t for better compatibility

* Switch size_t to uint32_t and uint64_t

* Clang-format

* Remove bad edits

+++ docs_generator/markdown_gen.hh
@@ -38,8 +38,8 @@ private:
             return "int32";
         } else if constexpr (std::is_same<T, uint32_t>()) {
             return "uint32";
-        } else if constexpr (std::is_same<T, size_t>()) {
-            return "size_t";
+        } else if constexpr (std::is_same<T, uint64_t>()) {
+            return "uint64";
         } else if constexpr (std::is_same<T, sp::angle_t>()) {
             return "float (degrees)";
         } else if constexpr (std::is_same<T, float>()) {
