+++ src/core/ecs/components/Renderable.cc
@@ -34,7 +34,7 @@ namespace ecs {
         if (dst.joints.empty()) dst.joints = src.joints;
     }
 
-    Renderable::Renderable(const std::string &modelName, size_t meshIndex)
+    Renderable::Renderable(const std::string &modelName, uint64_t meshIndex)
         : modelName(modelName), meshIndex(meshIndex) {
         if (modelName.empty()) {
             visibility = VisibilityMask::None;
